Remove test collector
