+++ crates/forge/Cargo.toml
@@ -15,7 +15,6 @@ indoc.workspace = true
 starknet_api.workspace = true
 tempfile.workspace = true
 rayon.workspace = true
-test-collector = { path = "../test-collector" }
 cheatnet = { path = "../cheatnet" }
 cairo-lang-runner.workspace = true
 cairo-lang-casm.workspace = true
