--- crates/test-collector/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-name = "test-collector"
-version = "1.0.0"
-edition = "2021"
-
-[features]
-testing = []
-
-[dependencies]
-rayon.workspace = true
-
-cairo-lang-casm.workspace = true
-cairo-lang-compiler.workspace = true
-cairo-lang-debug.workspace = true
-cairo-lang-defs.workspace = true
-cairo-lang-diagnostics.workspace = true
-cairo-lang-filesystem.workspace = true
-cairo-lang-lowering.workspace = true
-cairo-lang-test-runner.workspace = true
-cairo-lang-plugins.workspace = true
-cairo-lang-project.workspace = true
-cairo-lang-sierra-ap-change.workspace = true
-cairo-lang-sierra-gas.workspace = true
-cairo-lang-runner.workspace = true
-cairo-lang-semantic.workspace = true
-cairo-lang-sierra.workspace = true
-cairo-lang-sierra-generator.workspace = true
-cairo-lang-sierra-to-casm.workspace = true
-cairo-lang-starknet.workspace = true
-cairo-lang-syntax.workspace = true
-cairo-lang-utils.workspace = true
-anyhow.workspace = true
-cairo-vm.workspace = true
-cairo-felt.workspace = true
-itertools.workspace = true
-num-traits.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-thiserror.workspace = true
-smol_str = { version = "0.2.0", features = ["serde"] }
-num-bigint = "0.4"
-num-integer = "0.1"
-
-
-[lib]
-name = "test_collector"
